this is unfinished code\

+++ src/main/java/gg/sunken/shop/entity/trades/currency/ExpCurrencyCost.java
@@ -28,7 +28,7 @@ public class ExpCurrencyCost implements NpcCurrencyCost {
 
     @Override
     public boolean has(Player player) {
-        return economyProvider.has(player, (int) buyAmount);
+        return economyProvider.has(player, buyAmount);
     }
 
     @Override
